Remove unused tests
